[refactor] rename module
