+++ feature/login/src/main/AndroidManifest.xml
