+++ feature/login/src/main/java/com/mtc/login/LoginScreen.kt
@@ -1,7 +1,7 @@
-package com.mtc.signup
+package com.mtc.login
 
 import androidx.compose.runtime.Composable
 
 @Composable
-fun SignupRoute() {
+fun LoginRoute() {
 }
