+++ settings.gradle.kts
@@ -33,6 +33,6 @@ include(":core:buildconfig")
 include(":core:domain")
 include(":core:navigation")
 include(":feature:main")
-include(":feature:signup")
+include(":feature:login")
 include(":feature:home")
 include(":feature:addmatching")
